microservices/payments-service/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
prisma==0.11.0

# Cache & Queue
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
django-celery-beat==2.5.0

# Payment Gateways
stripe==7.8.0
paypalrestsdk==1.13.1

# Auth & Security
PyJWT==2.8.0
cryptography==41.0.7

# Utils
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2

# Server
gunicorn==21.2.0
uvicorn[standard]==0.25.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
factory-boy==3.3.0

# PDF Generation
reportlab==4.0.7

# Monitoring
sentry-sdk==1.39.1
